requirements.txt
langchain
langchain_experimental
chromadb==0.4.14
pandas
langchain_openai
lark
python-dotenv
faker
tabulate
huggingface_hub
transformers
sentence_transformers